requirements.txt
Django==2.2.3
django-model-utils==3.2.0
django-widget-tweaks==1.4.5
Markdown==3.1.1
Pygments==2.4.2
python-dotenv==0.10.3
pytz==2019.1
sqlparse==0.3.0
